Add logic to not to delete todo when todo id is not available in table

+++ src/main/java/com/crud/todo/service/TodoService.java
@@ -41,6 +41,8 @@ public class TodoService {
     }
 
     public boolean deleteTodoById(int todoId) {
+        Optional<Todo> todo = todoRepository.findById(todoId);
+        if (todo.isEmpty()) throw new TodoNotFoundException("Todo is not found!");
         todoRepository.deleteById(todoId);
         return true;
     }
